feat: Update CUDA container configuration and dependencies.

+++ config/docker/compose/cuda/setup.fish
@@ -1,2 +1,15 @@
 #!/usr/bin/env fish
-ln -s /mnt/usr/local/bin/* /usr/local/bin/
+if [ ! -e /usr/local/bin/micromamba ]
+    ln -s /mnt/usr/local/bin/* /usr/local/bin/
+end
+
+# 启动 ssh 服务
+cat ./passwd.usv | chpasswd
+mkdir -p /run/sshd
+chown root:root /run/sshd
+chmod 755 /run/sshd
+service ssh start
+
+# 启动 vnc 服务
+vncserver -kill ':*'
+nohup vncserver &>/var/log/vncserver.log &
